Convert from package to local
