+++ package.json
@@ -59,7 +59,6 @@
     "@casl/ability": "6.7.3",
     "@cdssnc/gcds-tokens": "1.20.1",
     "@gcforms/connectors": "workspace:*",
-    "@gcforms/drawer": "workspace:*",
     "@gcforms/types": "workspace:*",
     "@hcaptcha/react-hcaptcha": "^1.11.2",
     "@hcaptcha/types": "^1.0.4",
@@ -74,6 +73,7 @@
     "@types/mdx": "^2.0.13",
     "@zitadel/node": "2.0.37",
     "axios": "1.8.3",
+    "body-scroll-lock": "^4.0.0-beta.0",
     "clsx": "^2.1.1",
     "copy-to-clipboard": "3.3.3",
     "csv-writer": "^1.6.0",
@@ -110,7 +110,9 @@
     "react-loading-skeleton": "^3.4.0",
     "react-roving-tabindex": "^3.2.0",
     "react-select": "5.9.0",
+    "react-swipeable": "^7.0.2",
     "react-toastify": "11.0.5",
+    "react-transition-group": "^4.4.5",
     "reactflow": "^11.11.3",
     "server-only": "^0.0.1",
     "swr": "2.2.5",
@@ -150,6 +152,7 @@
     "@types/pg": "^8.10.9",
     "@types/react": "19.0.10",
     "@types/react-dom": "19.0.4",
+    "@types/react-transition-group": "^4",
     "@types/unorm": "^1",
     "@types/uuid": "^8.3.4",
     "autoprefixer": "^10.4.16",
