--- packages/drawer/package.json
@@ -1,36 +1,0 @@
-{
-  "name": "@gcforms/drawer",
-  "version": "0.1.1",
-  "description": "A React Component to build friendly dialogs that slide from the bottom",
-  "main": "dist",
-  "scripts": {
-    "prebuild": "rimraf dist",
-    "build": "tsc",
-    "prepack": "yarn build",
-    "drawer:prepack": "yarn prepack"
-  },
-  "license": "ISC",
-  "devDependencies": {
-    "@types/body-scroll-lock": "^2.6.1",
-    "@types/lodash.debounce": "^4.0.6",
-    "@types/react": "^19",
-    "@types/react-dom": "^19",
-    "@types/react-transition-group": "^4.2.4",
-    "parcel-bundler": "^1.12.4",
-    "react": "^19.0.0",
-    "react-dom": "^19.0.0",
-    "rimraf": "^3.0.2",
-    "typescript": "^5.5.2"
-  },
-  "dependencies": {
-    "body-scroll-lock": "^3.0.1",
-    "clsx": "^1.1.1",
-    "lodash.debounce": "^4.0.8",
-    "react-swipeable": "^7.0.2",
-    "react-transition-group": "^4.3.0"
-  },
-  "files": [
-    "dist"
-  ],
-  "packageManager": "yarn@4.6.0"
-}
